-[STORY] Remove the Bit library and replace it by a common Typescript library that compiles.
