--- neocom-infinity-shared/.bitmap
@@ -1,20 +1,0 @@
-/* THIS IS A BIT-AUTO-GENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. */
-
-/**
- * The Bitmap file is an auto generated file used by Bit to track all your Bit components. It maps the component to a folder in your file system.
- * This file should be committed to VCS(version control).
- * Components are listed using their component ID (https://bit.dev/reference/components/component-id).
- * If you want to delete components you can use the "bit remove <component-id>" command.
- * See the docs (https://bit.dev/reference/components/removing-components) for more information, or use "bit remove --help".
- */
-
-{
-    "neocom-domain": {
-        "name": "neocom-domain",
-        "scope": "innovative.neocom",
-        "version": "0.0.2",
-        "mainFile": "index.ts",
-        "rootDir": "neocom/neocom-domain"
-    },
-    "$schema-version": "17.0.0"
-}
