--- neocom-infinity-shared/.gitattributes
@@ -1,1 +1,0 @@
-.bitmap merge=bitmap-driver
