--- neocom-infinity-shared/README.md
@@ -1,14 +1,0 @@
-# BIT
-Bit is a library management and component framework that is being used to share Typescruipt code between the frontend and the backend.
-
-## Bit Commands
-3911  bit status
- 3912  bit link
- 3913  bit status
- 3914  bit install
- 3915  bit compile
- 3916  bit status
- 3917  bit export
- 3918  bit login
- 3919  bit login
- 3920  bit tag
