--- neocom-infinity-shared/neocom/neocom-domain/converter/NoActionConverter.ts
@@ -1,7 +1,0 @@
-import { IConverter } from "./IConverter.interface.js";
-
-export class NoActionConverter implements IConverter<any, any>{
-    convert(source: any): any {
-        return source
-    }
-}
