--- neocom-infinity-shared/neocom/neocom-domain/index.ts
@@ -1,5 +1,0 @@
-export { neocomDomain } from './neocom-domain.js'
-
-export { TypedResponseTransformer } from './transformer/TypedResponseTransformer.js'
-export { ITransformable } from './core/ITransformable.interface.js'
-export { IConverter}from './converter/IConverter.interface.js'
