--- neocom-infinity-shared/neocom/neocom-domain/neocom-domain.docs.mdx
@@ -1,11 +1,0 @@
----
-labels: ['NeocomDomain', 'library']
-description: 'Shared NeoCom classes and elements between the frontend and the backend. This module reduces duplicated classes on both developments.'
----
-
-A neocomDomain library that defines common data classes and core features compatible between Angular and NestJs.
-
-```ts
-neocomDomain()
-
-```
