--- neocom-infinity-shared/neocom/neocom-domain/neocom-domain.spec.ts
@@ -1,5 +1,0 @@
-import { neocomDomain } from './neocom-domain.js';
-
-it('renders with the correct text', () => {
-    expect(neocomDomain()).toEqual('NeoCom Domain available.');
-});
