--- neocom-infinity-shared/neocom/neocom-domain/neocom-domain.ts
@@ -1,3 +1,0 @@
-export function neocomDomain() {
-  return 'NeoCom Domain available.';
-}
