+++ neocom-infinity-shared/package.json
@@ -1,3 +1,25 @@
 {
-  "type": "module"
+    "name": "neocom-domain",
+    "version": "0.2.4",
+    "description": "Shared library for NeoCom Typesxcript code.",
+    "main": "dist/public-api.js",
+    "types": "dist/public-api.d.ts",
+    "type": "commonjs",
+    "scripts": {
+        "clean": "rm -rf ./dist",
+        "build": "npm run clean & tsc",
+        "test": "jest"
+    },
+    "keywords": [
+        "neocom",
+        "typescript"
+    ],
+    "author": "adamantinoo.git@gmail.com",
+    "license": "UNLICENSED",
+    "devDependencies": {
+        "@types/jest": "^29.5.12",
+        "jest": "^29.7.0",
+        "ts-jest": "^29.1.2",
+        "typescript": "^4.6.4"
+    }
 }
