+++ neocom-infinity-shared/src/converter/IConverter.interface.ts
