+++ neocom-infinity-shared/src/converter/NoActionConverter.ts
