+++ neocom-infinity-shared/src/core/ITransformable.interface.ts
