+++ neocom-infinity-shared/src/entities/V2.MiningOperation.ts
