--- neocom-shared/.vscode/extensions.json
@@ -1,4 +1,0 @@
-{
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=827846
-  "recommendations": ["angular.ng-template"]
-}
