--- neocom-shared/angular.json
@@ -1,52 +1,0 @@
-{
-  "$schema": "./node_modules/@angular/cli/lib/config/schema.json",
-  "cli": {
-    "analytics": "acc7c431-3559-4b47-be0d-90f66d01ac9c",
-    "defaultCollection": "@angular-eslint/schematics"
-  },
-  "version": 1,
-  "newProjectRoot": "projects",
-  "projects": {
-    "neocom-domain": {
-      "projectType": "library",
-      "root": "projects/neocom-domain",
-      "sourceRoot": "projects/neocom-domain/src",
-      "prefix": "lib",
-      "architect": {
-        "build": {
-          "builder": "@angular-devkit/build-angular:ng-packagr",
-          "options": {
-            "project": "projects/neocom-domain/ng-package.json"
-          },
-          "configurations": {
-            "production": {
-              "tsConfig": "projects/neocom-domain/tsconfig.lib.prod.json"
-            },
-            "development": {
-              "tsConfig": "projects/neocom-domain/tsconfig.lib.json"
-            }
-          },
-          "defaultConfiguration": "production"
-        },
-        "test": {
-          "builder": "@angular-devkit/build-angular:karma",
-          "options": {
-            "main": "projects/neocom-domain/src/test.ts",
-            "tsConfig": "projects/neocom-domain/tsconfig.spec.json",
-            "karmaConfig": "projects/neocom-domain/karma.conf.js"
-          }
-        },
-        "lint": {
-          "builder": "@angular-eslint/builder:lint",
-          "options": {
-            "lintFilePatterns": [
-              "projects/neocom-domain/**/*.ts",
-              "projects/neocom-domain/**/*.html"
-            ]
-          }
-        }
-      }
-    }
-  },
-  "defaultProject": "neocom-domain"
-}
