--- neocom-shared/projects/neocom-domain/README.md
@@ -1,24 +1,0 @@
-# NeocomDomain
-
-This library was generated with [Angular CLI](https://github.com/angular/angular-cli) version 13.3.0.
-
-## Code scaffolding
-
-Run `ng generate component component-name --project neocom-domain` to generate a new component. You can also use `ng generate directive|pipe|service|class|guard|interface|enum|module --project neocom-domain`.
-> Note: Don't forget to add `--project neocom-domain` or else it will be added to the default project in your `angular.json` file. 
-
-## Build
-
-Run `ng build neocom-domain` to build the project. The build artifacts will be stored in the `dist/` directory.
-
-## Publishing
-
-After building your library with `ng build neocom-domain`, go to the dist folder `cd dist/neocom-domain` and run `npm publish`.
-
-## Running unit tests
-
-Run `ng test neocom-domain` to execute the unit tests via [Karma](https://karma-runner.github.io).
-
-## Further help
-
-To get more help on the Angular CLI use `ng help` or go check out the [Angular CLI Overview and Command Reference](https://angular.io/cli) page.
