--- neocom-shared/projects/neocom-domain/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "neocom-domain",
-  "version": "0.0.1",
-  "peerDependencies": {
-    "@angular/common": "^13.3.0",
-    "@angular/core": "^13.3.0"
-  },
-  "dependencies": {
-    "tslib": "^2.3.0"
-  }
-}
