--- neocom-shared/projects/neocom-domain/src/converter/IConverter.interface.ts
@@ -1,3 +1,0 @@
-export interface IConverter<S, D> {
-    convert(source: S): D
-}
