--- neocom-shared/projects/neocom-domain/src/core/ITransformable.interface.ts
@@ -1,3 +1,0 @@
-export interface ITransformable<T>{
-    transform():T
-}
